Starting Gear Config - Started
I didn't want to use a separate block/TE for this, but I *really* didn't
want to mess with adding it to the GameController.

First attempt on a GUI with inventory.  What is sided-inventory?  I
might have removed some of that stuff from one of Tech's (fermenter)
TE's that I copied this from.  It crashes when you place the block.

Planning to add a button that copies your current inventory into the
upper inventory.  Nice dup "bug", but doesn't matter for
non-Survival-based game.  Need to make sure to include a copy of Armor.
Probably need 4 more slots in the GUI then.