Some Loot Block Fixes
Can right-click to change loot block state.
Can jump into block bottom face to change state.

Breaking a loot block crashes the game.  At that point, the block is now
Air and does not have the state that is trying to be gathered.  Probably
needs to be in the event handler?  Otherwise, how to stop the block
break from within the block's class?