Eureka !!!
Needed to send a message from client A(Attacker) to the server. and have
the server relay that message back to client B(target of attack ) and
have the movement of the player processed on the client B( target of
attack) to update its movement.   Apperntly somewhere MC is only taking
the Input from the client regarding its position :O